Add API key to calls and fix entities

+++ app/src/main/java/com/neaniesoft/myweather/data/model/entities/Wind.java
@@ -10,7 +10,7 @@ public class Wind {
     Double speed;
 
     @SerializedName("deg")
-    Integer directionDegrees;
+    Double directionDegrees;
 
     @SerializedName("gust")
     Double gustSpeed;
